Rework name assignment to values to better support renumbering (#90)
The recent work supporting loops adds complications to phi computations—these trickle into other areas of BlockBuilder, particularly with assignments of values to names. This PR removes names from the HIR values, using AssignHIR to associate names and values. I will provide a follow-up PR to analyze the flow of assignments to determine which need to be saved back out to the heap and what local variables need to be captured for use by inner blocks.