Preliminary work on saving named values (#91)
Hadron manipulates values exclusively in registers, so it needs some rules to determine what named values to save back to the heap after changing them. I've also added a few patches to clean up some parts of the CFG, specifically not connecting branches to continuing blocks if they contain a return inside the branch.

The NameSaver implementation highlights the need for block manipulation outside of BlockBuilder. I've decided to promote Block to a class and move some of the BlockBuilder code over to Block. That comes in the next patch.