Refactor block building to always load and store heap values (#93)
I realized my logic behind when to save a named value back to the heap was flawed. Because of the dynamic nature of SuperCollider dispatch, in combination with lexical scoping, the compiler can't be sure when some external code may modify named values. I have refactored the name resolution code to load and store named values in memory on every reference. This change simplifies the code and prioritizes correctness over speed. Eliminating redundant reads and writes will have to wait for an optimization PR later.