Move LIR appending to LinearFrame (#95)
This PR started a very large refactor to Hadron to support lowering code from HIR to LIR. I've stopped work on it here after moving the LIR append process to be a member function of LinearFrame. This work has felt "out on a limb" for a long time as I have been trying to get enough of the language interpreter working to do integration testing on the interpreter. It occurred to me today that I could build the Hadron debugger as an HIR interpreter, thus saving the complicated register allocation code work for after getting integration testing working. So I'm going to check this in and switch gears to focus on building a HIRTerpreter.