Compiling UnitTest and dependent classes (#97)
UnitTest support requires additional language features, including partial function application, global variables, and multiple assignment. I've removed string hash computation from the lexer as it was redundant with Symbol construction in the AST. The next step is to consider a new approach to the vistool, then start in on getting the unit tests to run.