Refactor the parser to use SCLang objects (#103)
This PR is part of the ongoing effort to make every artifact of the compilation pipeline accessible from the SC language itself. I've provided a new sclang file, HadronParseNode.sc, that defines the different objects in the parse tree and written C++ wrappers for each object. I then refactored the Bison to generate the new sclang data structures. Lastly, I've removed the old serialization code from the server and all the Parser unit tests. Once UnitTest is working, I'll revive the Parser testing on the language validation side, moving the code to SC.