Add a tool to dump parser JSON to stdout (#105)
This PR introduces the dump-diag binary, which can parse an input sclang file and print a parse tree to stdout. I've used a script to compute the percentage of files in the supercollider repository that successfully parses. I'll follow up with a separate PR to finish the parser and bring that number to 100%.