Expand Parser to parse most sclang code inside SuperCollider (#106)
To support using the parser in language tooling, I've gone through every file in the supercollider repository and added any missing functionality. This PR adds the following parser features:

 * Fix a crash on lexer errors
 * Parse list comprehensions (a.k.a. generators)
 * Lex the floating-point literals `pi` and `inf`
 * Parse the supported use cases for `pi` like `3pi` and `-pi`
 * Parse additional uses cases for `performList` syntax shortcuts using the `*` in argument lists
 * Add a few additional parses for `if` and `while` constructs
 * Fix a bug in the lexer when parsing strings that end in escape characters
 * Lex floating-point numbers in scientific notation, e.g. `1e8`
 * Lex radix floating-point and integer literals, e.g. `16rA` or `36r1.DEADBEEF`
 * Parse the `putSeries` syntax shortcut, e.g. `array[0,2..5] = "foo"`
 * Add adverb support to binops, e.g. `x <>>.in y`

There are a few files in the supercollider repository that still don't parse, but as far as I can tell they also fail to parse in legacy sclang, with one exception in `supercollider/testsuite/classlibrary/TestMethod.sc` which is testing that an input string should fail to compile, as it used to cause a crash in sclang when parsing.