Add Linux build support (#107)
This PR is a quick detour from my work moving the compilation artifacts over to sclang-accessible structures. The code needed only minor changes to compile with GCC, and I also had to add a few OS-specific functions to find the running binary path and map executable heap pages. Now that Hadron compiles on two operating systems, it's a priority to get the continuous integration setup (#30) to ensure that both builds remain healthy.