Add SCLang code to deserialize Hadron objects and visualize parse trees (#116)
Adds code to consume the output of the dump-diag JSON and convert it back to Hadron sclang data structures. When Hadron can run SCLang code well enough to generate and process these data structures in language, the deserialization from JSON may not be required. Still, we need it for now to access the data produced by the language runtime.

Then, it adds code to save a parse tree in DOT graph format for visualization.